Refactor commit: BlockContextualToolbar returns BlockToolbar directly
